+++ maths/Vector2.go
@@ -4,7 +4,7 @@ import "fmt"
 
 // Vector2 is a 2D vector
 type Vector2 struct {
-	X, Y float64
+	X, Y float32
 }
 
 func (v2 Vector2) Add(v Vector2) Vector2 {
@@ -23,15 +23,15 @@ func (v2 Vector2) Div(v Vector2) Vector2 {
 	return Vector2{v2.X / v.X, v2.Y / v.Y}
 }
 
-func (v2 Vector2) DistanceTo(v Vector2) float64 {
+func (v2 Vector2) DistanceTo(v Vector2) float32 {
 	return v2.Sub(v).Length()
 }
 
-func (v2 Vector2) Length() float64 {
+func (v2 Vector2) Length() float32 {
 	return (v2.X * v2.X) + (v2.Y * v2.Y)
 }
 
-func (v2 Vector2) LengthSquared() float64 {
+func (v2 Vector2) LengthSquared() float32 {
 	return v2.Length() * v2.Length()
 }
